backend/requirements.txt
CacheControl==0.12.5
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
coveralls==1.7.0
Django==2.2.1
django-rest-framework==0.1.0
djangorestframework==3.9.3
djangorestframework-simplejwt==4.3.0
docopt==0.6.2
firebase-admin==2.16.0
google-api-core==1.10.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==1.1.0
google-cloud-storage==1.15.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.10
grpcio==1.20.1
idna==2.8
msgpack==0.6.1
Pillow==6.0.0
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyJWT==1.7.1
pytz==2019.1
requests==2.21.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
